+++ U1-M5-Summative-Dion-Delcie/DelcieDionU1M5Summative/src/main/java/com/company/DelcieDionU1M5Summative/controller/PublisherController.java
@@ -23,20 +23,28 @@ public class PublisherController {
         return publisherDao.addPublisher(publisher);
     }
 
-    // ============ GET PUBLISHER ============
-
-    @RequestMapping(value = "/{id}",method = RequestMethod.GET)
-    @ResponseStatus(value = HttpStatus.OK)
-    public Publisher getAnPublisher(@PathVariable int id) {
-        return publisherDao.getPublisher(id);
-    }
-
     // ============ GET ALL PUBLISHERS ============
 
     @RequestMapping(method = RequestMethod.GET)
     @ResponseStatus(value = HttpStatus.OK)
     public List<Publisher> getAllPublishers() {
-        return publisherDao.getAllPublishers();
+        if (publisherDao.getAllPublishers().size() > 0) {
+            return publisherDao.getAllPublishers();
+        } else {
+            throw new IllegalArgumentException("No Publishers found.");
+        }
+    }
+
+    // ============ GET PUBLISHER ============
+
+    @RequestMapping(value = "/{id}",method = RequestMethod.GET)
+    @ResponseStatus(value = HttpStatus.OK)
+    public Publisher getPublisher(@PathVariable int id) {
+        if (publisherDao.getAllPublishers().contains(publisherDao.getPublisher(id))) {
+            return publisherDao.getPublisher(id);
+        } else {
+            throw new IllegalArgumentException("No Publisher with that ID found.");
+        }
     }
 
     // ============ UPDATE PUBLISHER ============
@@ -44,8 +52,12 @@ public class PublisherController {
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     @ResponseStatus(value = HttpStatus.OK)
     public void updatePublisher(@PathVariable int id, @RequestBody Publisher updatedPublisher) {
-        updatedPublisher.setId(id);
-        publisherDao.updatePublisher(updatedPublisher);
+        if (publisherDao.getAllPublishers().contains(publisherDao.getPublisher(id))) {
+            updatedPublisher.setId(id);
+            publisherDao.updatePublisher(updatedPublisher);
+        } else {
+            throw new IllegalArgumentException("Could not update, no publisher with matching ID.");
+        }
     }
 
     // ============ DELETE PUBLISHER ============
@@ -53,6 +65,10 @@ public class PublisherController {
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(value = HttpStatus.OK)
     public void deletePublisher(@PathVariable int id) {
-        publisherDao.deletePublisher(id);
+        if (publisherDao.getAllPublishers().contains(publisherDao.getPublisher(id))) {
+            publisherDao.deletePublisher(id);
+        } else {
+            throw new IllegalArgumentException("No Publisher with that ID found.");
+        }
     }
 }
